# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para gestionar ingresos y gastos personales",
        "title": "Financial Resume Engine API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/incomes": {
            "get": {
                "description": "Obtiene una lista de todos los ingresos del usuario",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incomes"
                ],
                "summary": "Listar ingresos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del usuario",
                        "name": "x-caller-id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/incomes.ListIncomesResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.UnauthorizedRequest"
                        }
                    }
                }
            },
            "post": {
                "description": "Crea un nuevo ingreso para el usuario",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incomes"
                ],
                "summary": "Crear un nuevo ingreso",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del usuario",
                        "name": "x-caller-id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Datos del ingreso",
                        "name": "income",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/incomes.CreateIncomeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/incomes.CreateIncomeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.BadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.UnauthorizedRequest"
                        }
                    }
                }
            }
        },
        "/api/v1/incomes/{id}": {
            "get": {
                "description": "Obtiene un ingreso por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incomes"
                ],
                "summary": "Obtener un ingreso específico",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del usuario",
                        "name": "x-caller-id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID del ingreso",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/incomes.GetIncomeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.BadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.UnauthorizedRequest"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.ResourceNotFound"
                        }
                    }
                }
            },
            "delete": {
                "description": "Elimina un ingreso existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incomes"
                ],
                "summary": "Eliminar un ingreso",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del usuario",
                        "name": "x-caller-id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID del ingreso",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.BadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.UnauthorizedRequest"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.ResourceNotFound"
                        }
                    }
                }
            },
            "patch": {
                "description": "Actualiza los datos de un ingreso existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incomes"
                ],
                "summary": "Actualizar un ingreso",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del usuario",
                        "name": "x-caller-id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID del ingreso",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos actualizados del ingreso",
                        "name": "income",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/incomes.UpdateIncomeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/incomes.UpdateIncomeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.BadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.UnauthorizedRequest"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.ResourceNotFound"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "errors.BadRequest": {
            "type": "object"
        },
        "errors.ResourceNotFound": {
            "type": "object"
        },
        "errors.UnauthorizedRequest": {
            "type": "object"
        },
...
        "incomes.CreateIncomeRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "incomes.CreateIncomeResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "incomes.GetIncomeResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "incomes.ListIncomesResponse": {
            "type": "object",
            "properties": {
                "incomes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/incomes.GetIncomeResponse"
                    }
                }
            }
        },
        "incomes.UpdateIncomeRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                }
            }
        },
        "incomes.UpdateIncomeResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    }
}